backend/requirements.txt
fastapi
uvicorn
deepface
tf-keras
opencv-python-headless
numpy
Pillow
mediapipe
scikit-image
python-multipart
